requirements.txt
astroid==2.3.3
audioread==2.1.8
Avpy==0.1.3
CacheControl==0.12.5
cachetools==3.1.1
certifi==2018.11.29
cffi==1.13.2
chardet==3.0.4
Click==7.0
cobs==1.1.3
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
ffmpeg==1.4
ffprobe==0.5
firebase-admin==3.2.0
Flask==1.1.1
Flask-Cors==3.0.8
gcloud==0.17.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==20.0.0
httplib2==0.14.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
jsonpickle==1.1
jwcrypto==0.6.0
jws==0.1.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
LEPL==5.1.3
librosa==0.7.2
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
msgpack==0.6.2
numba==0.47.0
numpy==1.17.4
oauth2client==3.0.0
pipenv==2018.11.26
profanity-check==1.0.3
pros-cli-v5==3.1.3
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.4
pydub==0.23.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyserial==3.4
python-dateutil==2.8.1
python-jwt==3.2.5
pytz==2019.3
pyzmq==17.1.2
raven==6.10.0
requests==2.21.0
requests-toolbelt==0.9.1
resampy==0.2.2
rfc6266-parser==0.0.5.post2
rsa==4.0
scan-build==2.0.15
scikit-learn==0.21.3
scipy==1.3.2
semantic-version==2.6.0
six==1.13.0
SoundFile==0.10.3.post1
tabulate==0.8.3
twilio==6.35.3
typed-ast==1.4.0
typing==3.6.6
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.7.8
virtualenv-clone==0.5.3
Werkzeug==0.16.0
wrapt==1.11.2
